requirements.txt
Flask
Flask-SQLalchemy
Pandas
reportlab
Gunicorn
python-dotenv
supabase
supabase-py